+++ src/main/java/de/caritas/cob/userservice/api/facade/assignsession/AssignSessionFacade.java
@@ -1,7 +1,5 @@
 package de.caritas.cob.userservice.api.facade.assignsession;
 
-import static java.util.concurrent.CompletableFuture.supplyAsync;
-
 import de.caritas.cob.userservice.api.admin.service.rocketchat.RocketChatRemoveFromGroupOperationService;
 import de.caritas.cob.userservice.api.facade.EmailNotificationFacade;
 import de.caritas.cob.userservice.api.facade.RocketChatFacade;
@@ -18,7 +16,6 @@ import de.caritas.cob.userservice.api.service.statistics.event.AssignSessionStat
 import de.caritas.cob.userservice.statisticsservice.generated.web.model.UserRole;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Supplier;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -58,7 +55,7 @@ public class AssignSessionFacade {
 
     updateSessionInDatabase(session, consultant);
     addNewConsultantToRocketChatGroup(session, consultant);
-    supplyAsync(removeUnauthorizedMembersFromGroups(session, consultant));
+    removeUnauthorizedMembersFromGroups(session, consultant);
     sendEmailForConsultantChange(session, consultant);
 
     statisticsService.fireEvent(
@@ -83,19 +80,15 @@ public class AssignSessionFacade {
     rocketChatFacade.removeSystemMessagesFromRocketChatGroup(rcGroupId);
   }
 
-  private Supplier<Void> removeUnauthorizedMembersFromGroups(Session session,
-      Consultant consultant) {
-    return () -> {
-      var memberList = rocketChatFacade.retrieveRocketChatMembers(session.getGroupId());
-      removeUnauthorizedMembersFromGroup(session, consultant, memberList);
+  private void removeUnauthorizedMembersFromGroups(Session session, Consultant consultant) {
+    var memberList = rocketChatFacade.retrieveRocketChatMembers(session.getGroupId());
+    removeUnauthorizedMembersFromGroup(session, consultant, memberList);
 
-      if (session.hasFeedbackChat()) {
-        var feedbackMemberList =
-            rocketChatFacade.retrieveRocketChatMembers(session.getFeedbackGroupId());
-        removeUnauthorizedMembersFromFeedbackGroup(session, consultant, feedbackMemberList);
-      }
-      return null;
-    };
+    if (session.hasFeedbackChat()) {
+      var feedbackMemberList = rocketChatFacade
+          .retrieveRocketChatMembers(session.getFeedbackGroupId());
+      removeUnauthorizedMembersFromFeedbackGroup(session, consultant, feedbackMemberList);
+    }
   }
 
   private void removeUnauthorizedMembersFromGroup(Session session, Consultant consultant,